feat: move formatUrlWithWorkspaceId to core/public/utils
