+++ src/core/public/index.ts
@@ -357,4 +357,4 @@ export {
 
 export { WORKSPACE_TYPE } from '../utils';
 
-export { getWorkspaceIdFromUrl } from './utils';
+export { getWorkspaceIdFromUrl, formatUrlWithWorkspaceId } from './utils';
