+++ src/core/public/utils/index.ts
@@ -31,5 +31,5 @@
 export { shareWeakReplay } from './share_weak_replay';
 export { Sha256 } from './crypto';
 export { MountWrapper, mountReactNode } from './mount';
-export { getWorkspaceIdFromUrl } from './workspace';
+export { getWorkspaceIdFromUrl, formatUrlWithWorkspaceId } from './workspace';
 export { WORKSPACE_PATH_PREFIX, WORKSPACE_TYPE } from '../../utils';
