+++ src/core/public/utils/workspace.test.ts
@@ -3,7 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { getWorkspaceIdFromUrl } from './workspace';
+import { getWorkspaceIdFromUrl, formatUrlWithWorkspaceId } from './workspace';
+import { httpServiceMock } from '../mocks';
 
 describe('#getWorkspaceIdFromUrl', () => {
   it('return workspace when there is a match', () => {
@@ -12,5 +13,20 @@ describe('#getWorkspaceIdFromUrl', () => {
 
   it('return empty when there is not a match', () => {
     expect(getWorkspaceIdFromUrl('http://localhost/w2/foo')).toEqual('');
+  });
+});
+
+describe('#formatUrlWithWorkspaceId', () => {
+  const basePathWithoutWorkspaceBasePath = httpServiceMock.createSetupContract().basePath;
+  it('return url with workspace prefix when format with a id provided', () => {
+    expect(
+      formatUrlWithWorkspaceId('/app/dashboard', 'foo', basePathWithoutWorkspaceBasePath)
+    ).toEqual('http://localhost/w/foo/app/dashboard');
+  });
+
+  it('return url without workspace prefix when format without a id', () => {
+    expect(
+      formatUrlWithWorkspaceId('/w/foo/app/dashboard', '', basePathWithoutWorkspaceBasePath)
+    ).toEqual('http://localhost/app/dashboard');
   });
 });
